+++ gems.rb
@@ -6,6 +6,7 @@
 source "https://rubygems.org"
 
 gem "falcon"
+gem "async-grpc-xds", github: "socketry/async-grpc-xds", ref: "fd9867c9155550982fff6eb77ed75a98c84fda49"
 
 gemspec